Rename common to libs to avoid naming conflicts
