+++ libs/device.py
