+++ libs/libvhd.py
