+++ libs/libvhd_shelve.py
