+++ libs/poolhelper.py
