+++ libs/tapdisk.py
