+++ overlay/etc/xapi.d/plugins/suspend-resume-datapath
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import XenAPIPlugin
-from xapi.storage.datapath import dmsetup, losetup, tapdisk, image
+from xapi.storage.libs import dmsetup, losetup, tapdisk, image
 from xapi.storage import log
